requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Flask==1.1.2
geographiclib==1.50
geopy==2.0.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
LivePopularTimes==1.0
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
requests==2.23.0
urllib3==1.25.9
Werkzeug==1.0.1
python-dotenv==0.14.0